Sync TOC: Scroll to the active TOC item if needed
After sync, make sure the activated TOC item is in the viewport.
Otherwise, and if the TOC is a separate scroll area, scroll item into
viewport.

Do not use "scrollIntoView()" since this does not work in Chrome in this
case and "scrollIntoViewIfNeeded()" since it is non-standard and not
supported by Firefox and other browsers.