Sync TOC: In some cases automatically collapse siblings on the path
When a TOC item is selected other than via clicking on it (e.g. when in
a topic clicking on a link), automatically collapse all siblings on the
path (the siblings of the TOC item and all siblings of the ancestors).
This is based on the assumption that when navigating in the TOC tree it
is not wanted that items are automatically collapsed, but in other cases
it is wanted (because when something is automatically expanded, it
should also be automatically collapsed again).

In addition, mark the ancestor items of the selected TOC items with the
class "selected-p".