Enable deep linking
Implemented as follows:
The hash part of the URL of the main page represent the location of the
content page of the <iframe>. When navigating to another topic, the hash
part has to be updated (the web browser history should be modified
without adding new states). When the main page is loaded, the content
page <iframe> is set accordingly, if the URL has a hash part containing
a deep link.