Search results page filter: Improve automatic (un)selection of ancestors
In the search page filter tree, a checkbox should only be (un)checked
automatically on (un)checking all sub-checkboxes if and only if the
result number is equal to and not less than the sum of the results of
the sub-checkboxes.

Please note that (un)checking a checkbox always (un)checks all
descendant checkboxes since this is probably what is wanted in most
cases. To see results of a chapter only, without the results of its
subchapters, first select the chapter and then deselect all its
subchapters.