New concept: Change TOC auto scroll behavior to show more above+below
In the TOC sidebar, when scrolling automatically to show an item in the
visible area, show the TOC item in the upper third instead of scroll as
less as possible resulting the item being displayed either at the very
top or at the very bottom.

The upper third seems to me to be the better choice than centering it,
because the parent/ancestors require(s) less space than the
children/descendants.