New concept: Add book scope synchronization
When switching to another book, the book scope in the search field
should updated as well (when no scope is set, nothing should happen).

Add an option to specify whether initially no scope or the scope should
be set to the current book. This option should be remembered by a
cookie.