New concept: Implement solution for navigating having deep links
Going back in the browser history can cause issues in combination with
deep linking, since the navigation is done in the iframe except for the
search page which is not shown in the iframe. The problem is that when
going back to a search page, the search might need to be submitted again
and for this the query must be known.

Ways that don't work:

- Deep link containing query as hash of the top window (...#q=...) set
via `history.pushState(...)`: top window hash might not be restored when
navigation happens also in the content iframe

- Query as hash or as query of the content iframe set via
`history.pushState(...)`: conflicts with existing hashes/queries of
content pages and does not work with external content pages

- Data URL used in content iframe containing the query: not allowed in
Internet Explorer for security reasons


Chosen solution:

-> For full search set content iframe instead of doing a remote request
and get result from live DOM of the iframe when loaded