For small height, search-as-you-type proposal list should be scrollable
On a mobile device with a small screen, when typing a search query most
of the height will be covered by the on screen keyboard which reduces
the available height. In such cases the search-as-you-type proposal list
should be scrollable vertically in order to be able to access all
proposals.