Implement Bookmarks (embedded mode only)
In contrast to the legacy UI, show the bookmarks in the main area
(instead of a topic or instead of search results), not as a side bar.

Support bookmarking of search links and the ability to change the label
of a bookmark when creating it (in the legacy UI neither of these is
supported).