Bookmarks: Correctly calculate links (use relative links where possible)
Including a workaround for a bug of the legacy UI: non-topic links (e.g.
".../nav/..." instead of ".../topic/...") are stored absolutely instead
of relatively (which doesn't work, because by default a random port
number is chosen when restarting Eclipse). In such cases, bookmark a
link like e.g. ".../nav/..." as a relative link starting with
"/../nav/".