Make sure bookmarks also work in Internet Explorer
The legacy bookmarks HTML pages that are used in the new UI contain
<meta http-equiv="Expires" content="-1">
But since they are not displayed as HTML pages, but only loaded by
JavaScript via remote request and then parsed as text, this will be
ignored. To avoid getting stale content for the same URL, add a
timestamp to the URL as a workaround to avoid caching.