Show progress when indexing; be aware that a 'g' RegExp is not stateless
In a full search the progress of creating the search index should be
displayed until the search results will be displayed.

In JavaScript a regular expression with the global flag 'g' is not
stateless for the same string. So make sure to reset it when reusing it
e.g. via "myRegexp.lastIndex = 0;".